test(content-system): tighten assertions and relocate the di wiring pin
StoredTreeTest pinned instance identity on an untargeted root with assertNotSame, which the mutation aliasing contract declares non-normative; the value-preservation assertEquals stays as the pin. PropertyTypeTest compared toSchema()['translatable'] against translatable(), two reads of one object that could drift together, and covered the number declaration of admits() only for an integer value; the flag tests now assert literals for both states and the provider gains the float and rejection rows. The DI definition pin in the style-parity test needs no booted kernel, so it moves to the unit twin beside the ContentRouteCompilerPassTest precedent, leaving the integration file the behavioral pin that does. Three @see references gain the short-name-plus-import form the docblock rule requires.
