+++ tests/integration/Core/Framework/ContentSystem/Storefront/ContentLayoutStorefrontRenderTest.php
@@ -12,6 +12,7 @@ use Shopware\Core\Framework\Test\TestCaseBase\IntegrationTestBehaviour;
 use Shopware\Core\Test\Stub\Framework\IdsCollection;
 use Shopware\Storefront\Test\Controller\StorefrontControllerTestBehaviour;
 use Shopware\Tests\Integration\Core\Framework\ContentSystem\ContentLayoutFixtureBehaviour;
+use Shopware\Tests\Integration\Core\Framework\ContentSystem\Validation\MediaImageWriteGateTest;
 use Symfony\Component\HttpFoundation\Response;
 
 /**
@@ -95,7 +96,7 @@ class ContentLayoutStorefrontRenderTest extends TestCase
 
     /**
      * `Sw:Media:Image.html.twig:47` branches on `media is not null`; the write gate only requires `mediaId` to
-     * be filled, not that the media row exists (see {@see \Shopware\Tests\Integration\Core\Framework\ContentSystem\Validation\MediaImageWriteGateTest::testPersistsMediaImageWriteWithFilledMediaId()}),
+     * be filled, not that the media row exists (see {@see MediaImageWriteGateTest::testPersistsMediaImageWriteWithFilledMediaId()}),
      * so a layout wired to a dangling media id persists and this is the leg that renders it.
      */
     #[TestDox('renders the placeholder markup instead of an img tag when the element\'s media id names no media row')]
