+++ tests/unit/Core/Framework/ContentSystem/Layout/Type/Specification/PropertyTypeTest.php
@@ -92,6 +92,11 @@ class PropertyTypeTest extends TestCase
             StoredValue::ofInt(3),
         ];
 
+        yield 'float on a lone number declaration' => [
+            new PropertyType('number', false, null, null),
+            StoredValue::ofFloat(3.5),
+        ];
+
         yield 'integer on an all-primitive union carrying integer' => [
             new PropertyType(['string', 'integer'], false, null, null),
             StoredValue::ofInt(3),
@@ -179,6 +184,11 @@ class PropertyTypeTest extends TestCase
             StoredValue::ofFloat(3.5),
         ];
 
+        yield 'string on a lone number declaration' => [
+            new PropertyType('number', false, null, null),
+            StoredValue::ofString('3.5'),
+        ];
+
         yield 'boolean on an all-primitive union of string and integer' => [
             new PropertyType(['string', 'integer'], false, null, null),
             StoredValue::ofBool(true),
@@ -258,11 +268,21 @@ class PropertyTypeTest extends TestCase
         ];
     }
 
-    #[TestDox('reads the translatable flag the published schema carries')]
-    public function testTranslatableReadsTheFlagThePublishedSchemaCarries(): void
+    #[TestDox('reads the translatable flag a translatable declaration carries')]
+    public function testTranslatableReadsTheFlagATranslatableDeclarationCarries(): void
     {
         $type = new PropertyType('string', true, null, null);
 
-        static::assertSame($type->toSchema()['translatable'], $type->translatable());
+        static::assertTrue($type->translatable());
+        static::assertTrue($type->toSchema()['translatable']);
+    }
+
+    #[TestDox('reports no translatable flag for a non-translatable declaration')]
+    public function testTranslatableReportsNoFlagForANonTranslatableDeclaration(): void
+    {
+        $type = new PropertyType('string', false, null, null);
+
+        static::assertFalse($type->translatable());
+        static::assertFalse($type->toSchema()['translatable']);
     }
 }
